+++ frontend/src/app/components/forms/Form.js
@@ -1,0 +1,44 @@
+import React from "react";
+import SignUpOptions from './SignUpOptions';
+import ManualSignUp from './ManualSignUp';
+import StudentVerifyEmail from './StudentVerifyEmail';
+import StudentClassCode from './StudentClassCode';
+import TeacherExposureToUs from './TeacherExposureToUs';
+import SubscriptionType from './SubscriptionType';
+import PaymentPage from './PaymentPage';
+import Enterprise from './Enterprise';
+
+export default function Form(props) {
+    const { stage } = props;
+    if (stage.type === 'student') {
+        switch (stage.step) {
+            case 1:
+                return <SignUpOptions {...props} />
+            case 2:
+                return <ManualSignUp {...props} />
+            case 3:
+                return <StudentVerifyEmail {...props} />
+            case 4:
+                return <StudentClassCode {...props} />
+            default:
+                return <SignUpOptions {...props} />
+        }
+    }
+    else if (stage.type === 'teacher') {
+        switch (stage.step) {
+            case 1:
+                return <SignUpOptions {...props} />
+            case 2:
+                return <ManualSignUp {...props} />
+            case 3:
+                return <TeacherExposureToUs {...props} />
+            case 4:
+                return <Enterprise {...props} />
+            case 5:
+                return <PaymentPage {...props} />
+            default:
+                return <SignUpOptions {...props} />
+        }
+    }
+    return <SignUpOptions />
+}
