+++ frontend/src/app/components/forms/ManualSignUp.js
@@ -1,0 +1,78 @@
+import React from "react"
+
+export default function ManualSignUp(props) {
+    return (
+        <div>
+            <div className="row">
+                <div className="col-12">
+                    <div className="form-floating">
+                        <input
+                            placeholder="Email Address"
+                            value={props.email}
+                            data-state="email"
+                            onChange={props.handleChange}
+                            type="email"
+                            className="form-control"
+                            id="email"
+                            aria-describedby="emailHelp" />
+                        <label htmlFor="email" className="form-label">Email Address</label>
+                    </div>
+                </div>
+                <div className="col-lg-6 col-12">
+                    <div className="form-floating mt-3">
+                        <input
+                            placeholder="Password"
+                            value={props.password}
+                            onFocus={props.handleFocus}
+                            data-state="password"
+                            onChange={props.handleChange}
+                            type="password"
+                            className="form-control"
+                            id="password" />
+                        <label style={{ color: "black" }} htmlFor="password" className="form-label">Password</label>
+                    </div>
+                </div>
+                <div className="col-lg-6 col-12">
+                    <div className="form-floating mt-3">
+                        <input
+                            placeholder="Re-enter Password"
+                            value={props.verifiedPassword}
+                            onFocus={props.handleFocus}
+                            data-state="verifiedPassword"
+                            onChange={props.handleChange}
+                            type="password"
+                            className="form-control"
+                            id="verifiedPassword" />
+                        <label htmlFor="verifiedPassword" className="form-label">Re-enter Password</label>
+                    </div>
+                </div>
+                <div className="col-12 mt-3">
+                    <div className="form-check">
+                        <input
+                            value={props.rememberMe}
+                            data-state="rememberMe"
+                            onChange={props.handleChange}
+                            type="checkbox"
+                            className="form-check-input"
+                            id="rememberMe" />
+                        <label className="form-check-label" htmlFor="rememberMe">Remember Me</label>
+                    </div>
+                </div>
+            </div>
+            <div className="row mt-3 d-flex justify-content-around">
+                <button
+                    className="col-5 btn btn-secondary"
+                    onClick={() => { props.setStage({ step: 1, type: 'student' }) }}
+                >
+                    Back
+                </button>
+                <button
+                    className="col-5 btn btn-primary"
+                    onClick={() => { props.setStage({ step: 3, type: 'student' }) }}
+                >
+                    Continue
+                </button>
+            </div>
+        </div>
+    )
+}
